requirements.txt
wheel
aiogram==2.19
aiohttp==3.8.1
aioschedule==0.5.2
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
easyocr==1.4.1
frozenlist==1.3.0
idna==3.3
imageio==2.16.1
multidict==6.0.2
networkx==2.7.1
numpy==1.22.3
opencv-python-headless==4.5.4.60
packaging==21.3
pyparsing==3.0.7
python-bidi==0.4.2
python-dotenv==0.19.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qrcode==7.3.1
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
soupsieve==2.3.1
tifffile==2022.3.16
torch==1.11.0
torchvision==0.12.0
typing-extensions==4.1.1
urllib3==1.26.9
yarl==1.7.2
